consolidates planning and readme files
