--- SERIAL_MONITOR_TROUBLESHOOTING.md
@@ -1,124 +1,0 @@
-# Arduino Serial Monitor Troubleshooting Guide
-
-## Problem: No messages appearing in Serial Monitor
-
-### Step 1: Upload the Test Sketch
-1. Upload `arduino_serial_test.ino` to your Arduino
-2. This simple sketch will help diagnose the issue
-
-### Step 2: Check Serial Monitor Settings
-
-#### In Arduino IDE:
-1. **Open Serial Monitor**: Tools → Serial Monitor (or Ctrl+Shift+M)
-2. **Set Baud Rate**: Make sure it's set to **9600** (bottom right of Serial Monitor)
-3. **Check Port**: Tools → Port → Select the correct COM port (e.g., COM3, COM4, /dev/ttyUSB0)
-
-### Step 3: Common Issues & Solutions
-
-#### Issue 1: Wrong Baud Rate
-- **Problem**: Serial Monitor baud rate doesn't match code
-- **Solution**: Set Serial Monitor to 9600 baud (bottom right dropdown)
-
-#### Issue 2: Wrong COM Port
-- **Problem**: Arduino connected to different port than selected
-- **Solution**: 
-  - Disconnect Arduino USB
-  - Check Tools → Port menu
-  - Reconnect Arduino USB
-  - Select the new port that appears
-
-#### Issue 3: Board Not Recognized
-- **Problem**: Computer doesn't recognize Arduino
-- **Solution**:
-  - Try different USB cable
-  - Try different USB port
-  - Install Arduino drivers if needed
-  - Check if LED on Arduino is lit (indicates power)
-
-#### Issue 4: Serial Monitor Not Opening
-- **Problem**: Serial Monitor window doesn't open
-- **Solution**:
-  - Close all Serial Monitor windows
-  - Try Tools → Serial Monitor again
-  - Restart Arduino IDE
-
-#### Issue 5: Board Resets When Opening Serial Monitor
-- **Problem**: Arduino resets when Serial Monitor opens
-- **Solution**: This is normal for some boards - wait a moment for messages
-
-### Step 4: Test Different Baud Rates
-
-If 9600 doesn't work, try these common rates:
-- 115200
-- 57600
-- 38400
-- 19200
-- 9600
-
-Update the code and Serial Monitor to match.
-
-### Step 5: Hardware Check
-
-#### Visual Indicators:
-- **Power LED**: Should be lit on Arduino
-- **TX/RX LEDs**: Should blink when sending data
-- **Built-in LED**: Should blink every 500ms (from test sketch)
-
-#### If No LEDs Blink:
-- Arduino not running code
-- Wrong board selected in Tools → Board
-- Upload failed (check for error messages)
-
-### Step 6: Alternative Testing
-
-#### Try This Simple Code:
-```cpp
-void setup() {
-  Serial.begin(9600);
-  Serial.println("Hello World!");
-}
-
-void loop() {
-  delay(1000);
-  Serial.println("Still working...");
-}
-```
-
-### Step 7: Platform-Specific Issues
-
-#### Windows:
-- Check Device Manager for COM port conflicts
-- Try running Arduino IDE as Administrator
-- Install CH340 or FTDI drivers if using clone boards
-
-#### Mac:
-- Check System Preferences → Security & Privacy for blocked drivers
-- Try different USB port
-- Check Console app for error messages
-
-#### Linux:
-- Add user to dialout group: `sudo usermod -a -G dialout $USER`
-- Check permissions: `ls -l /dev/ttyUSB*` or `/dev/ttyACM*`
-
-### Expected Output
-
-When working correctly, you should see:
-```
-=== ARDUINO SERIAL TEST ===
-If you can see this message, Serial is working!
-Baud rate: 9600
-Board type: [your board type]
-==========================
-Heartbeat: 2000 ms
-Heartbeat: 4000 ms
-Heartbeat: 6000 ms
-```
-
-### Still Not Working?
-
-1. **Try a different Arduino board** if available
-2. **Check Arduino IDE version** - try updating or downgrading
-3. **Try different computer** to isolate hardware vs software issue
-4. **Check Arduino board type** in Tools → Board - make sure it matches your hardware
-
-The built-in LED should blink every 500ms regardless of Serial communication - if it's not blinking, the issue is with the Arduino itself, not Serial communication.
