+++ server/env.example
@@ -1,10 +1,7 @@
-# Database Configuration
-DB_USER=postgres
-DB_HOST=localhost
-DB_NAME=spy_database
-DB_PASSWORD=your_password_here
-DB_PORT=5432
+# Neon Database Configuration
+# Get your connection string from the Neon dashboard: https://console.neon.tech
+# Format: postgresql://username:password@host/database?sslmode=require
+POSTGRES_URL=postgresql://username:password@host/database?sslmode=require
 
-# Server Configuration
-PORT=3001
-NODE_ENV=development 
+# Alternative: You can also use DATABASE_URL (Neon sometimes provides this)
+# DATABASE_URL=postgresql://username:password@host/database?sslmode=require
