+++ server/package.json
@@ -2,16 +2,17 @@
   "name": "spy-database-server",
   "version": "1.0.0",
   "description": "Express server for spy database application",
+  "type": "module",
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
     "dev": "nodemon server.js",
     "setup-db": "node setup.js",
     "setup-db-sql": "psql -U postgres -f database.sql",
-    "setup-db-vercel": "node setup-vercel.js"
+    "setup-db-neon": "node setup-neon.js"
   },
   "dependencies": {
-    "@vercel/postgres": "^0.5.1",
+    "@neondatabase/serverless": "^1.0.2",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
